QPID-1447 : Add additional testing for null TopicDelete Policy Configuration and fix for NPEs.
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@951161 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/broker-plugins/experimental/slowconsumerdetection/src/main/java/org/apache/qpid/server/virtualhost/plugin/policies/TopicDeletePolicy.java
@@ -93,7 +93,7 @@ public class TopicDeletePolicy implements SlowConsumerPolicyPlugin
                                  "Consuming to slow.");
 
             // Actively delete non autoDelete queues if deletePersistent is set
-            if (!q.isAutoDelete() && _configuration.deletePersistent())
+            if (!q.isAutoDelete() && (_configuration != null && _configuration.deletePersistent()))
             {
                 CurrentActor.get().message(q.getLogSubject(), TopicDeletePolicyMessages.TDP_DELETING_QUEUE());
                 q.delete();
